grosse modif

+++ include/server/File.hpp
@@ -5,7 +5,7 @@
 // Login   <antoine.plaskowski@epitech.eu>
 // 
 // Started on  Tue Dec 22 23:15:52 2015 Antoine Plaskowski
-// Last update Sat Dec 26 14:43:43 2015 Antoine Plaskowski
+// Last update Sat Dec 26 15:36:56 2015 Antoine Plaskowski
 //
 
 #ifndef		FILE_HPP_
@@ -23,10 +23,10 @@ public:
   std::string const	&get_name(void) const;
   uint8_t const	*get_data(void) const;
   uintmax_t	get_size(void) const;
-  uintmax_t	get_id(void) const;
+  uint8_t	get_id(void) const;
   std::list<File *> const	&get_files(void) const;
 private:
-  static uintmax_t	m_id_max;
+  static uint8_t	m_id_max;
   static std::list<File *>	m_files;
   std::string const	m_name;
   uint8_t	*m_data;
